Add files via upload
The program works;
Small stuff to fix

+++ include/push_swap.h
@@ -30,7 +30,7 @@
 
 typedef struct s_list
 {
-	int					content;
+	long					content;
 	int					cost;
 	int					index;
 	struct s_list		*bestfriend;
